Create CoreData logic for product, order item and order. Create Cart viewController, presenter and coordinator.
(cherry picked from commit 9b388c815607aa0e7f95117076c42becce62703e)

+++ SimpleMarket/SimpleMarket/Models/Product/Product+Decodable.swift
@@ -16,6 +16,6 @@ struct DecodableProduct: Decodable {
     let price: Double
     let image: String
     let description: String
-    let stock: Double?
+    let stock: Int?
     let offer: Double?
 }
